Fix linter

+++ src/client/terrain/util/chunk.rs
@@ -2,9 +2,9 @@ use crate::prelude::*;
 
 pub const CHUNK_SIZE: usize = 32;
 pub const PADDED_CHUNK_SIZE: usize = CHUNK_SIZE + 2;
-pub const PADDED_CHUNK_USIZE: usize = PADDED_CHUNK_SIZE as usize;
+pub const PADDED_CHUNK_USIZE: usize = PADDED_CHUNK_SIZE;
 pub const CHUNK_LENGTH: usize =
-    (PADDED_CHUNK_SIZE * PADDED_CHUNK_SIZE * PADDED_CHUNK_SIZE) as usize;
+    PADDED_CHUNK_SIZE * PADDED_CHUNK_SIZE * PADDED_CHUNK_SIZE;
 
 pub struct Chunk {
     pub data: [BlockId; CHUNK_LENGTH],
@@ -23,7 +23,6 @@ impl Chunk {
         self.get_unpadded(x + 1, y + 1, z + 1)
     }
 
-
     pub fn get_unpadded(&self, x: usize, y: usize, z: usize) -> BlockId {
         self.data[Self::index(x, y, z)]
     }
@@ -39,7 +38,7 @@ impl Chunk {
 
     #[rustfmt::skip]
     pub fn index(x: usize, y: usize, z: usize) -> usize {
-      if (x >= PADDED_CHUNK_SIZE as usize) || (y >= PADDED_CHUNK_SIZE as usize) || (z >= PADDED_CHUNK_SIZE as usize) {
+      if (x >= PADDED_CHUNK_SIZE) || (y >= PADDED_CHUNK_SIZE) || (z >= PADDED_CHUNK_SIZE) {
         panic!("Index out of bounds: ({}, {}, {})", x, y, z);
       }
         x + PADDED_CHUNK_USIZE * (y + PADDED_CHUNK_USIZE * z)
